split ValueObjectRemoverDocBlockRector
