+++ tests/Rector/DomainDrivenDesign/ValueObjectRemoverDocBlockRector/Correct/correct.php.inc
@@ -1,0 +1,20 @@
+<?php
+
+namespace Rector\Tests\Rector\DomainDrivenDesign\ValueObjectRemoverDocBlockRector;
+
+use Rector\Tests\Rector\DomainDrivenDesign\ValueObjectRemoverDocBlockRector\Source\SomeChildOfValueObject;
+
+class ActionClass
+{
+    /**
+     * @var string|null
+     */
+    private $someChildValueObject;
+
+    public function someFunction()
+    {
+        $this->someChildValueObject = new SomeChildOfValueObject('value');
+
+        $someChildValueObject = new SomeChildOfValueObject();
+    }
+}
