+++ tests/Rector/DomainDrivenDesign/ValueObjectRemoverDocBlockRector/Source/SomeChildOfValueObject.php
@@ -1,0 +1,7 @@
+<?php declare(strict_types=1);
+
+namespace Rector\Tests\Rector\DomainDrivenDesign\ValueObjectRemoverDocBlockRector\Source;
+
+class SomeChildOfValueObject extends SomeValueObject
+{
+}
