+++ tests/Rector/DomainDrivenDesign/ValueObjectRemoverDocBlockRector/Source/SomeValueObject.php
@@ -1,0 +1,8 @@
+<?php declare(strict_types=1);
+
+namespace Rector\Tests\Rector\DomainDrivenDesign\ValueObjectRemoverDocBlockRector\Source;
+
+class SomeValueObject
+{
+
+}
