+++ tests/Rector/DomainDrivenDesign/ValueObjectRemoverDocBlockRector/ValueObjectRemoverDocBlockRectorTest.php
@@ -1,15 +1,14 @@
 <?php declare(strict_types=1);
 
-namespace Rector\Tests\Rector\DomainDrivenDesign\ValueObjectRemoverRector;
+namespace Rector\Tests\Rector\DomainDrivenDesign\ValueObjectRemoverDocBlockRector;
 
 use Iterator;
 use Rector\Testing\PHPUnit\AbstractRectorTestCase;
 
 /**
- * @covers \Rector\Rector\DomainDrivenDesign\ValueObjectRemover\ValueObjectRemoverRector
  * @covers \Rector\Rector\DomainDrivenDesign\ValueObjectRemover\ValueObjectRemoverDocBlockRector
  */
-final class ValueObjectRemoverRectorTest extends AbstractRectorTestCase
+final class ValueObjectRemoverDocBlockRectorTest extends AbstractRectorTestCase
 {
     /**
      * @dataProvider provideWrongToFixedFiles()
@@ -24,7 +23,6 @@ final class ValueObjectRemoverRectorTest extends AbstractRectorTestCase
         yield [__DIR__ . '/Wrong/wrong.php.inc', __DIR__ . '/Correct/correct.php.inc'];
         yield [__DIR__ . '/Wrong/wrong2.php.inc', __DIR__ . '/Correct/correct2.php.inc'];
         yield [__DIR__ . '/Wrong/wrong3.php.inc', __DIR__ . '/Correct/correct3.php.inc'];
-        yield [__DIR__ . '/Wrong/wrong4.php.inc', __DIR__ . '/Correct/correct4.php.inc'];
     }
 
     protected function provideConfig(): string
