+++ tests/Rector/DomainDrivenDesign/ValueObjectRemoverDocBlockRector/Wrong/wrong.php.inc
@@ -1,0 +1,20 @@
+<?php
+
+namespace Rector\Tests\Rector\DomainDrivenDesign\ValueObjectRemoverDocBlockRector;
+
+use Rector\Tests\Rector\DomainDrivenDesign\ValueObjectRemoverDocBlockRector\Source\SomeChildOfValueObject;
+
+class ActionClass
+{
+    /**
+     * @var SomeChildOfValueObject|null
+     */
+    private $someChildValueObject;
+
+    public function someFunction()
+    {
+        $this->someChildValueObject = new SomeChildOfValueObject('value');
+
+        $someChildValueObject = new SomeChildOfValueObject();
+    }
+}
