+++ tests/Rector/DomainDrivenDesign/ValueObjectRemoverDocBlockRector/config.yml
@@ -1,0 +1,4 @@
+services:
+    Rector\Rector\DomainDrivenDesign\ValueObjectRemover\ValueObjectRemoverDocBlockRector:
+        $valueObjectsToSimpleTypes:
+            'Rector\Tests\Rector\DomainDrivenDesign\ValueObjectRemoverDocBlockRector\Source\SomeValueObject': 'string'
